week18 done

+++ week18/creatn.cpp
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-long long arr[1002];
+int arr[1002];
 
 int main(){
     ios::sync_with_stdio(0);
@@ -9,9 +9,9 @@ int main(){
     cout.tie(0);
     freopen("creatn.inp", "r", stdin);
     freopen("creatn.out", "w", stdout);
-    long long n, sum = 0;
+    int n, sum = 0;
     cin >> n;
-    for(long long i = 0; i < 3*n; ++i){
+    for(int i = 0; i < 3*n; ++i){
         cin >> arr[i];
     }
     sort(arr, arr + (3*n));
